requirements.txt
alembic==1.0.6
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-Misaka==1.0.0
Flask-SQLAlchemy==2.3.2
Flask-Login==0.4.1
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.0.7
MarkupSafe==1.1.0
misaka==2.1.1
psycopg2==2.7.3.2
pycparser==2.19
python-dotenv==0.10.1
python-editor==1.0.3
requests==2.21.0
six==1.12.0
SQLAlchemy==1.3.0
urllib3==1.24.1
Werkzeug==0.14.1
oauthlib==3.0.1
pyOpenSSL==19.0.0

